Playing with currying, chaining and underscores
A super hacky, quick implementation of some ideas from

https://discourse.julialang.org/t/fixing-the-piping-chaining-issue

Parse chains of `/>` and `\>` at the rough same precedence as `|>`, but
treat them as a "frontfix/backfix operator" for function calls such that
the succeeding function call becomes curried in first or last argument.
Thus, the following

    x  />  f(y)  \>  g(z)

is parsed as

    (chain x (/> f y) (\> g z))

and lowered to the equivalent of

    chain(x, a->f(a, y), b->g(z, b))

Also add lowering of underscore as strictly tight-binding placeholder
syntax. (Super hacky - more forms should be allowed! This is just for
experimentation).